refactor core module and introduce new ExtractorResults class
